requirements.txt
numpy
flask
torch
torchaudio
torchvision
Pillow
matplotlib